+++ src/main/java/roomescape/auth/TokenProvider.java
@@ -9,6 +9,7 @@ import java.util.Date;
 import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import roomescape.auth.exception.AuthorizationException;
 
 @Component
 public class TokenProvider {
