+++ src/main/java/roomescape/auth/exception/AccessDeniedException.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.exception;
 
 public class AccessDeniedException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "잘못된 접근입니다.";
