+++ src/main/java/roomescape/auth/exception/AuthorizationException.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.exception;
 
 public class AuthorizationException extends RuntimeException {
     public AuthorizationException(String message) {
