+++ src/main/java/roomescape/global/advice/GlobalExceptionHandler.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.auth.AccessDeniedException;
-import roomescape.auth.AuthorizationException;
+import roomescape.auth.exception.AccessDeniedException;
+import roomescape.auth.exception.AuthorizationException;
 import roomescape.global.dto.ErrorResponse;
 
 @RestControllerAdvice
